<build>: make cmake file to compile all
